type error fiddling

+++ train.py
@@ -140,7 +140,7 @@ class VLMExtractor(BaseFeaturesExtractor):
         with torch.no_grad():
             raw_features = self.agent.embed(obs, max_batch=self.vlm_internal_batch_size)
         features = self.proj(raw_features)
-        return features
+        return features.float()
 
 # --- Setup Function ---
 def setup_model_and_env(config):
